requirements.txt
streamlit==1.16.0
plotly==5.11.0
Pillow==9.4.0
pandas==2.0.1